// src/models/Scheme.ts
export interface Scheme {
	primary: string;
	surfaceTint: string;
	onPrimary: string;
	primaryContainer: string;
	onPrimaryContainer: string;
	secondary: string;
	onSecondary: string;
	secondaryContainer: string;
	onSecondaryContainer: string;
	tertiary: string;
	onTertiary: string;
	tertiaryContainer: string;
	onTertiaryContainer: string;
	error: string;
	onError: string;
	errorContainer: string;
	onErrorContainer: string;
	background: string;
	onBackground: string;
	surface: string;
	onSurface: string;
	surfaceVariant: string;
	onSurfaceVariant: string;
	outline: string;
	outlineVariant: string;
	shadow: string;
	scrim: string;
	inverseSurface: string;
	inverseOnSurface: string;
	inversePrimary: string;
	primaryFixed: string;
	onPrimaryFixed: string;
	primaryFixedDim: string;
	onPrimaryFixedVariant: string;
	secondaryFixed: string;
	onSecondaryFixed: string;
	secondaryFixedDim: string;
	onSecondaryFixedVariant: string;
	tertiaryFixed: string;
	onTertiaryFixed: string;
	tertiaryFixedDim: string;
	onTertiaryFixedVariant: string;
	surfaceDim: string;
	surfaceBright: string;
	surfaceContainerLowest: string;
	surfaceContainerLow: string;
	surfaceContainer: string;
	surfaceContainerHigh: string;
	surfaceContainerHighest: string;
}